fedimint_mint_client::backup
Module recovery

Structs

CompressedBlindedMessage 🔒

Newtype over BlindedMessage to enable Ord

EcashRecoveryFinalState

MintRecovery

MintRecoveryStateV0

The state machine used for fast-forwarding backup from point when it was taken to the present time by following epoch history items from the time the snapshot was taken.

Enums

MintRecoveryState