Struct fedimint_mint_client::backup::recovery::MintRestoreInProgressState

pub(crate) struct MintRestoreInProgressState {
    start_epoch: u64,
    next_epoch: u64,
    end_epoch: u64,
    spendable_note_by_nonce: BTreeMap<Nonce, (Amount, SpendableNote)>,
    pending_outputs: BTreeMap<OutPoint, (TieredMulti<(CompressedBlindedMessage, Option<NoteIssuanceRequest>)>, BTreeMap<PeerId, Vec<BlindedSignatureShare>>)>,
    pending_nonces: BTreeMap<CompressedBlindedMessage, (NoteIssuanceRequest, NoteIndex, Amount)>,
    next_pending_note_idx: Tiered<NoteIndex>,
    last_mined_nonce_idx: Tiered<NoteIndex>,
    threshold: u64,
    pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>,
    tbs_pks: Tiered<AggregatePublicKey>,
    gap_limit: u64,
}

The state machine used for fast-forwarding backup from point when it was taken to the present time by following epoch history items from the time the snapshot was taken.

The caller is responsible for creating it, and then feeding it in order all valid consensus items from the epoch history between time taken (or even somewhat before it) and present time.

Fields
start_epoch: u64next_epoch: u64end_epoch: u64spendable_note_by_nonce: BTreeMap<Nonce, (Amount, SpendableNote)>

Nonces that we track that are currently spendable.

pending_outputs: BTreeMap<OutPoint, (TieredMulti<(CompressedBlindedMessage, Option<NoteIssuanceRequest>)>, BTreeMap<PeerId, Vec<BlindedSignatureShare>>)>

Outputs (by OutPoint) we track federation member confirmations for blind nonces.

Once we get enough confirmation (valid shares), these become new spendable notes.

Note that NoteIssuanceRequest is optional, as sometimes we might need to handle a tx where only some of the blind nonces were in the pool. A None means that this blind nonce/message is there only for validation purposes, and will actually not create a spendable_note_by_nonce

pending_nonces: BTreeMap<CompressedBlindedMessage, (NoteIssuanceRequest, NoteIndex, Amount)>

Next nonces that we expect might soon get used. Once we see them, we move the tracking to pending_outputs

Note: since looking up nonces is going to be the most common operation the pool is kept shared (so only one lookup is enough), and replenishment is done each time a note is consumed.

next_pending_note_idx: Tiered<NoteIndex>

Tail of pending. pending_notes is filled by generating note with this index and incrementing it.

last_mined_nonce_idx: Tiered<NoteIndex>

LastECashNoteIndex but tracked in flight. Basically max index of any note that got a partial sig from the federation (initialled from the backup value). TODO: One could imagine a case where the note was issued but not get any partial sigs yet. Very unlikely in real life scenario, but worth considering.

threshold: u64

Threshold

pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>

Public key shares for each peer

Used to validate contributed consensus items

tbs_pks: Tiered<AggregatePublicKey>

Aggregate public key for each amount tier

gap_limit: u64

The number of nonces we look-ahead when looking for mints (per each amount).

Implementations

impl MintRestoreInProgressState

fn transitions( &self, operation_id: OperationId, context: &MintClientContext, global_context: &DynGlobalClientContext ) -> Vec<StateTransition<MintRestoreStateMachine>>

async fn make_progress<'a>( self, api: DynGlobalApi, decoders: ModuleDecoderRegistry, epoch_pk: PublicKey, secret: DerivableSecret ) -> Self

fn fetch_epochs_stream<'a>( api: DynGlobalApi, epoch_pk: PublicKey, decoders: ModuleDecoderRegistry, epoch_range: Range<u64> ) -> impl Stream<Item = (u64, SignedEpochOutcome)> + 'a

Fetch epochs in a given range and send them over sender

Since WASM’s spawn does not support join handles, we indicate errors via sender itself.

TODO: could be internal to recovery_loop?

impl MintRestoreInProgressState

pub fn from_backup( current_epoch_count: u64, backup: EcashBackup, gap_limit: u64, tbs_pks: Tiered<AggregatePublicKey>, pub_key_shares: BTreeMap<PeerId, Tiered<PublicKeyShare>>, secret: &DerivableSecret ) -> Self

fn fill_initial_pending_nonces( &mut self, amount: Amount, secret: &DerivableSecret )

Fill each tier pool to the gap limit

fn add_next_pending_nonce_in_pending_pool( &mut self, amount: Amount, secret: &DerivableSecret )

Add next nonce from amount tier to the next_pending_note_idx

pub fn handle_input(&mut self, input: &MintInput)

pub fn handle_output( &mut self, out_point: OutPoint, output: &MintOutput, secret: &DerivableSecret )

fn observe_nonce_idx_being_used( &mut self, amount: Amount, note_idx: NoteIndex, secret: &DerivableSecret )

React to a valid pending nonce being tracked being used in the epoch history

(Possibly) increment the self.last_mined_nonce_idx, then replenish the pending pool to always maintain at least gap_limit of pending nonces in each amount tier.

pub fn handle_output_confirmation( &mut self, peer_id: PeerId, sigs: &MintConsensusItem )

pub fn is_done(&self) -> bool

pub(crate) fn handle_consensus_item( &mut self, peer_id: PeerId, item: &ConsensusItem, processed_txs: &mut HashSet<TransactionId>, rejected_txs: &BTreeSet<TransactionId>, secret: &DerivableSecret )

fn finalize(self) -> EcashRecoveryFinalState