fedimint_mint_client

Function represent_amount

pub fn represent_amount<K>(
    amount: Amount,
    current_denominations: &TieredCounts,
    tiers: &Tiered<K>,
    denomination_sets: u16,
    fee_consensus: &FeeConsensus,
) -> TieredCounts

Determines the denominations to use when representing an amount

Algorithm tries to leave the user with a target number of denomination_sets starting at the lowest denomination. self gives the denominations that the user already has.