fedimint_mint_client

Module output

State machines for mint outputs

Structs

MintOutputCommon

MintOutputCommonV0

MintOutputStateMachine

MintOutputStateMachineV0

MintOutputStatesAborted

See MintOutputStates

MintOutputStatesCreated

See MintOutputStates

MintOutputStatesCreatedMulti

See MintOutputStates

MintOutputStatesFailed

See MintOutputStates

MintOutputStatesSucceeded

See MintOutputStates

NoteIssuanceRequest

Keeps the data to generate SpendableNote once the mint successfully processed the transaction signing the corresponding BlindNonce.

Enums

MintOutputStates

State machine managing the e-cash issuance process related to a mint output.

Constants

BLINDING_KEY_CHILD_ID 🔒

Child ID used to derive the blinding key from a note’s DerivableSecret

SPEND_KEY_CHILD_ID 🔒

Child ID used to derive the spend key from a note’s DerivableSecret

Functions

verify_blind_share

Panics