fedimint_mint_client

Struct MintClientModule

pub struct MintClientModule {
    pub(crate) federation_id: FederationId,
    pub(crate) cfg: MintClientConfig,
    pub(crate) secret: DerivableSecret,
    pub(crate) secp: Secp256k1<All>,
    pub(crate) notifier: ModuleNotifier<MintClientStateMachines>,
    pub client_ctx: ClientContext<Self>,
    pub(crate) balance_update_sender: Sender<()>,
}

The MintClientModule is responsible for handling e-cash minting operations. It interacts with the mint server to issue, reissue, and validate e-cash notes.

Derivable Secret

The DerivableSecret is a cryptographic secret that can be used to derive other secrets. In the context of the MintClientModule, it is used to derive the blinding and spend keys for e-cash notes. The DerivableSecret is initialized when the MintClientModule is created and is kept private within the module.

Blinding Key

The blinding key is derived from the DerivableSecret and is used to blind the e-cash note during the issuance process. This ensures that the mint server cannot link the e-cash note to the client that requested it, providing privacy for the client.

Spend Key

The spend key is also derived from the DerivableSecret and is used to spend the e-cash note. Only the client that possesses the DerivableSecret can derive the correct spend key to spend the e-cash note. This ensures that only the owner of the e-cash note can spend it.

Fields
federation_id: FederationIdcfg: MintClientConfigsecret: DerivableSecretsecp: Secp256k1<All>notifier: ModuleNotifier<MintClientStateMachines>client_ctx: ClientContext<Self>balance_update_sender: Sender<()>
Implementations

impl MintClientModule

pub async fn prepare_plaintext_ecash_backup( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> Result<EcashBackup>

impl MintClientModule

pub async fn try_repair_wallet(&self, gap_limit: u64) -> Result<RepairSummary>

Attempts to fix inconsistent wallet states. Breaks privacy guarantees and is destructive!

Invalid states that are fixable this way:

Already-spent e-cash being in the wallet

E-cash nonces that would be used to issue new notes already being used

When invalid notes are found, they are removed from the wallet. Make sure that the user has a backup of their seed before running this function.

async fn gap_till_next_nonce_used( &self, module_api: &DynModuleApi, amount: Amount, base_index: u64, gap_limit: u64, ) -> Result<Option<u64>>

Checks up to gap_limit nonces starting from base_index for having being used already.

If the nonce at base_index is used, returns Some(0), if it’s unused returns None. If there’s an unused nonce and then a used one returns Some(1).

impl MintClientModule

pub(crate) async fn create_sufficient_input( &self, dbtx: &mut DatabaseTransaction<'_>, min_amount: Amount, ) -> Result<Vec<(ClientInput<MintInput>, SpendableNote)>>

pub async fn get_notes_tier_counts( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> TieredCounts

👎Deprecated since 0.5.0: Use get_note_counts_by_denomination instead

Returns the number of held e-cash notes per denomination

pub async fn get_available_notes_by_tier_counts( &self, dbtx: &mut DatabaseTransaction<'_>, counts: TieredCounts, ) -> (TieredMulti<SpendableNoteUndecoded>, TieredCounts)

Pick SpendableNotes by given counts, when available

Return the notes picked, and counts of notes that were not available.

pub async fn create_output( &self, dbtx: &mut DatabaseTransaction<'_>, operation_id: OperationId, notes_per_denomination: u16, exact_amount: Amount, ) -> ClientOutputBundle<MintOutput, MintClientStateMachines>

Creates a mint output close to the given amount, issuing e-cash notes such that the client holds notes_per_denomination notes of each e-cash note denomination held.

pub async fn get_note_counts_by_denomination( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> TieredCounts

Returns the number of held e-cash notes per denomination

pub async fn get_wallet_summary( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> TieredCounts

👎Deprecated since 0.5.0: Use get_note_counts_by_denomination instead

Returns the number of held e-cash notes per denomination

pub async fn estimate_spend_all_fees(&self) -> Amount

Estimates the total fees to spend all currently held notes.

This is useful for calculating max withdrawable amounts, where all notes will be spent. Notes that are uneconomical to spend (fee >= value) are excluded from the calculation since the wallet won’t spend them.

pub async fn await_output_finalized( &self, operation_id: OperationId, out_point: OutPoint, ) -> Result<()>

Wait for the e-cash notes to be retrieved. If this is not possible because another terminal state was reached an error describing the failure is returned.

pub async fn consolidate_notes( &self, dbtx: &mut DatabaseTransaction<'_>, ) -> Result<Vec<(ClientInput<MintInput>, SpendableNote)>>

Provisional implementation of note consolidation

When a certain denomination crosses the threshold of notes allowed, spend some chunk of them as inputs.

Return notes and the sume of their amount.

pub fn create_input_from_notes( &self, notes: TieredMulti<SpendableNote>, ) -> Result<Vec<(ClientInput<MintInput>, SpendableNote)>>

Create a mint input from external, potentially untrusted notes

pub(crate) async fn spend_notes_oob( &self, dbtx: &mut DatabaseTransaction<'_>, notes_selector: &impl NotesSelector, amount: Amount, try_cancel_after: Duration, ) -> Result<(OperationId, Vec<MintClientStateMachines>, TieredMulti<SpendableNote>)>

pub async fn await_spend_oob_refund( &self, operation_id: OperationId, ) -> SpendOOBRefund

pub(crate) async fn select_notes( dbtx: &mut DatabaseTransaction<'_>, notes_selector: &impl NotesSelector, requested_amount: Amount, fee_consensus: FeeConsensus, ) -> Result<TieredMulti<SpendableNote>>

Select notes with requested_amount using notes_selector.

pub(crate) async fn get_all_spendable_notes( dbtx: &mut DatabaseTransaction<'_>, ) -> TieredMulti<SpendableNoteUndecoded>

pub(crate) async fn get_next_note_index( &self, dbtx: &mut DatabaseTransaction<'_>, amount: Amount, ) -> NoteIndex

pub fn new_note_secret_static( secret: &DerivableSecret, amount: Amount, note_idx: NoteIndex, ) -> DerivableSecret

Derive the note DerivableSecret from the Mint’s secret the amount tier and note_idx

Static to help re-use in other places, that don’t have a whole Self available

E-Cash Note Creation

When creating an e-cash note, the MintClientModule first derives the blinding and spend keys from the DerivableSecret. It then creates a NoteIssuanceRequest containing the blinded spend key and sends it to the mint server. The mint server signs the blinded spend key and returns it to the client. The client can then unblind the signed spend key to obtain the e-cash note, which can be spent using the spend key.

pub(crate) async fn new_note_secret( &self, amount: Amount, dbtx: &mut DatabaseTransaction<'_>, ) -> DerivableSecret

We always keep track of an incrementing index in the database and use it as part of the derivation path for the note secret. This ensures that we never reuse the same note secret twice.

pub async fn new_ecash_note( &self, amount: Amount, dbtx: &mut DatabaseTransaction<'_>, ) -> (NoteIssuanceRequest, BlindNonce)

pub async fn reissue_external_notes<M: Serialize + Send>( &self, oob_notes: OOBNotes, extra_meta: M, ) -> Result<OperationId>

Try to reissue e-cash notes received from a third party to receive them in our wallet. The progress and outcome can be observed using MintClientModule::subscribe_reissue_external_notes. Can return error of type ReissueExternalNotesError

pub async fn subscribe_reissue_external_notes( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<ReissueExternalNotesState>>

Subscribe to updates on the progress of a reissue operation started with MintClientModule::reissue_external_notes.

pub async fn spend_notes<M: Serialize + Send>( &self, min_amount: Amount, try_cancel_after: Duration, include_invite: bool, extra_meta: M, ) -> Result<(OperationId, OOBNotes)>

👎Deprecated since 0.5.0: Use spend_notes_with_selector instead, with SelectNotesWithAtleastAmount to maintain the same behavior

Fetches and removes notes of at least amount min_amount from the wallet to be sent to the recipient out of band. These spends can be canceled by calling MintClientModule::try_cancel_spend_notes as long as the recipient hasn’t reissued the e-cash notes themselves yet.

The client will also automatically attempt to cancel the operation after try_cancel_after time has passed. This is a safety mechanism to avoid users forgetting about failed out-of-band transactions. The timeout should be chosen such that the recipient (who is potentially offline at the time of receiving the e-cash notes) had a reasonable timeframe to come online and reissue the notes themselves.

pub async fn spend_notes_with_selector<M: Serialize + Send>( &self, notes_selector: &impl NotesSelector, requested_amount: Amount, try_cancel_after: Duration, include_invite: bool, extra_meta: M, ) -> Result<(OperationId, OOBNotes)>

Fetches and removes notes from the wallet to be sent to the recipient out of band. The note selection algorithm is determined by note_selector. See the NotesSelector trait for available implementations.

These spends can be canceled by calling MintClientModule::try_cancel_spend_notes as long as the recipient hasn’t reissued the e-cash notes themselves yet.

The client will also automatically attempt to cancel the operation after try_cancel_after time has passed. This is a safety mechanism to avoid users forgetting about failed out-of-band transactions. The timeout should be chosen such that the recipient (who is potentially offline at the time of receiving the e-cash notes) had a reasonable timeframe to come online and reissue the notes themselves.

pub async fn send_oob_notes<M: Serialize + Send>( &self, amount: Amount, extra_meta: M, ) -> Result<OOBNotes>

Send e-cash notes for the requested amount.

When this method removes ecash notes from the local database it will do so atomically with creating a SendPaymentEvent that contains the notes in out of band serilaized from. Hence it is critical for the integrator to display this event to ensure the user always has access to his funds.

This method operates in two modes:

Offline mode: If exact notes are available in the wallet, they are spent immediately without contacting the federation. A SendPaymentEvent is emitted and the notes are returned.

Online mode: If exact notes are not available, the method contacts the federation to trigger a reissuance transaction to obtain the proper denominations. The method will block until the reissuance completes, at which point a SendPaymentEvent is emitted and the notes are returned.

If the method enters online mode and is cancelled, e.g. the future is dropped, before the reissue transaction is confirmed, any reissued notes will be returned to the wallet and we do not emit a SendPaymentEvent.

Before selection of the ecash notes the amount is rounded up to the nearest multiple of 512 msat.

pub(crate) async fn try_spend_exact_notes_dbtx( &self, dbtx: &mut DatabaseTransaction<'_>, amount: Amount, federation_id: FederationId, extra_meta: Value, ) -> Option<OOBNotes>

Try to spend exact notes from the current balance. Returns Some(OOBNotes) if exact notes are available, None otherwise.

pub fn validate_notes(&self, oob_notes: &OOBNotes) -> Result<Amount>

Validate the given notes and return the total amount of the notes. Validation checks that:

the federation ID is correct

the note has a valid signature

the spend key is correct.

pub async fn check_note_spent(&self, oob_notes: &OOBNotes) -> Result<bool>

Contacts the mint and checks if the supplied notes were already spent.

Caution: This reduces privacy and can lead to race conditions. DO NOT rely on it for receiving funds unless you really know what you are doing.

pub async fn try_cancel_spend_notes(&self, operation_id: OperationId)

Try to cancel a spend operation started with MintClientModule::spend_notes_with_selector. If the e-cash notes have already been spent this operation will fail which can be observed using MintClientModule::subscribe_spend_notes.

pub async fn subscribe_spend_notes( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<SpendOOBState>>

Subscribe to updates on the progress of a raw e-cash spend operation started with MintClientModule::spend_notes_with_selector.

pub(crate) async fn mint_operation( &self, operation_id: OperationId, ) -> Result<OperationLogEntry>

pub(crate) async fn delete_spendable_note( client_ctx: &ClientContext<MintClientModule>, dbtx: &mut DatabaseTransaction<'_>, amount: Amount, note: &SpendableNote, )

pub async fn advance_note_idx(&self, amount: Amount) -> Result<DerivableSecret>

pub async fn reused_note_secrets( &self, ) -> Vec<(Amount, NoteIssuanceRequest, BlindNonce)>

Returns secrets for the note indices that were reused by previous clients with same client secret.