Trait fedimint_mint_client::MintClientExt

pub trait MintClientExt {
    fn reissue_external_notes<'life0, 'async_trait, M>(
        &'life0 self,
        notes: TieredMulti<SpendableNote>,
        extra_meta: M
    ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>
    where
        M: 'async_trait + Serialize + Send,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_reissue_external_notes<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, ReissueExternalNotesState>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn spend_notes<'life0, 'async_trait, M>(
        &'life0 self,
        min_amount: Amount,
        try_cancel_after: Duration,
        extra_meta: M
    ) -> Pin<Box<dyn Future<Output = Result<(OperationId, TieredMulti<SpendableNote>)>> + Send + 'async_trait>>
    where
        M: 'async_trait + Serialize + Send,
        Self: 'async_trait,
        'life0: 'async_trait;
    fn try_cancel_spend_notes<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_spend_notes<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, SpendOOBState>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn await_restore_finished<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Required Methods

fn reissue_external_notes<'life0, 'async_trait, M>(
&'life0 self,
notes: TieredMulti<SpendableNote>,
extra_meta: M
) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>where
M: 'async_trait + Serialize + Send,
Self: 'async_trait,
'life0: 'async_trait,

Try to reissue e-cash notes received from a third party to receive them in our wallet. The progress and outcome can be observed using MintClientExt::subscribe_reissue_external_notes.

fn subscribe_reissue_external_notes<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, ReissueExternalNotesState>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Subscribe to updates on the progress of a reissue operation started with MintClientExt::reissue_external_notes.

fn spend_notes<'life0, 'async_trait, M>(
&'life0 self,
min_amount: Amount,
try_cancel_after: Duration,
extra_meta: M
) -> Pin<Box<dyn Future<Output = Result<(OperationId, TieredMulti<SpendableNote>)>> + Send + 'async_trait>>where
M: 'async_trait + Serialize + Send,
Self: 'async_trait,
'life0: 'async_trait,

Fetches and removes notes of at least amount min_amount from the wallet to be sent to the recipient out of band. These spends can be canceled by calling MintClientExt::try_cancel_spend_notes as long as the recipient hasn’t reissued the e-cash notes themselves yet.

The client will also automatically attempt to cancel the operation after try_cancel_after time has passed. This is a safety mechanism to avoid users forgetting about failed out-of-band transactions. The timeout should be chosen such that the recipient (who is potentially offline at the time of receiving the e-cash notes) had a reasonable timeframe to come online and reissue the notes themselves.

fn try_cancel_spend_notes<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Try to cancel a spend operation started with MintClientExt::spend_notes. If the e-cash notes have already been spent this operation will fail which can be observed using MintClientExt::subscribe_spend_notes.

fn subscribe_spend_notes<'life0, 'async_trait>(
&'life0 self,
operation_id: OperationId
) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<'_, SpendOOBState>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Subscribe to updates on the progress of a raw e-cash spend operation started with MintClientExt::spend_notes.

fn await_restore_finished<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Awaits the backup restoration to complete