Crate fedimint_mint_client

Re-exports

pub use fedimint_mint_common as common;

Modules

common

config

Structs

BackupRequest

BlindNonce

Nonce but blinded by the user key

InsufficientBalanceError

MintClientContext

MintClientGen

MintClientModule

MintCommonGen

MintConsensusItem

Data structures taking into account different amount tiers A consenus item from one of the federation members contributing partials signatures to blind nonces submitted in it

MintInput

MintMeta

MintModuleTypes

MintOutput

MintOutputBlindSignatures

Result of Federation members confirming MintOutput by contributing partial signatures via MintConsensusItem

MintOutputOutcome

MintOutputSignatureShare

Blind signature share from one Federation peer for a single MintOutput

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

NoteIndex

An index used to deterministically derive Notes

SignedBackupRequest

SpendOOBRefund

SpendableNote

A Note with associated secret key that allows to proof ownership (spend it)

Enums

CombineError

MintClientStateMachines

MintError

PeerErrorType

ReissueExternalNotesState

The high-level state of a reissue operation started with MintClientExt::reissue_external_notes.

SpendOOBState

The high-level state of a raw e-cash spend operation started with MintClientExt::spend_notes.

Constants

DEFAULT_MAX_NOTES_PER_DENOMINATION

By default, the maximum notes per denomination when change-making for users

KIND

LOG_TARGET

Traits

MintClientExt

Functions

parse_ecash

select_notes_from_stream