Crate fedimint_mint_client

Re-exports

pub use fedimint_mint_common as common;

Modules

backup

client_db

Database keys used throughout the mint client module

config

db

input 🔒

State machines for mint inputs

oob 🔒

State machines for out-of-band transmitted e-cash notes

output

State machines for mint outputs

Structs

BackupRequest

BlindNonce

Nonce but blinded by the user key

InsufficientBalanceError

MintClientContext

MintClientInit

MintClientModule

The MintClientModule is responsible for handling e-cash minting operations. It interacts with the mint server to issue, reissue, and validate e-cash notes.

MintCommonInit

MintInputV0

MintModuleTypes

MintOperationMeta

MintOutputBlindSignature

Result of Federation members confirming MintOutput by contributing partial signatures via MintConsensusItem

MintOutputOutcomeV0

MintOutputV0

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

NoteIndex

An index used to deterministically derive Notes

OOBNotes

An encapsulation of FederationId and e-cash notes in the form of TieredMulti<SpendableNote> for the purpose of spending e-cash out-of-band. Also used for validating and reissuing such out-of-band notes.

OOBReissueTag 🔒

OOBSpendTag 🔒

SelectNotesWithAtleastAmount

Select notes with total amount of at least request_amount. If more than requested amount of notes are returned it was because exact change couldn’t be made, and the next smallest amount will be returned.

SelectNotesWithExactAmount

Select notes with total amount of exactly request_amount. If the amount cannot be represented with the available denominations an error is returned, this does not mean that the balance is too low.

SignedBackupRequest

SpendOOBRefund

SpendableNote

A Note with associated secret key that allows to proof ownership (spend it)

UnknownMintInputVariantError

UnknownMintOutputOutcomeVariantError

UnknownMintOutputVariantError

Enums

MintClientStateMachines

MintConsensusItem

The mint module currently doesn’t define any consensus items and generally throws an error on encountering one. To allow old clients to still decode blocks in the future, should we decide to add consensus items, this has to be an enum with only a default variant.

MintInput

MintInputError

MintOperationMetaVariant

MintOutput

MintOutputError

MintOutputOutcome

OOBNotesData 🔒

ReissueExternalNotesState

The high-level state of a reissue operation started with MintClientModule::reissue_external_notes.

SpendOOBState

The high-level state of a raw e-cash spend operation started with MintClientModule::spend_notes.

Constants

DEFAULT_MAX_NOTES_PER_DENOMINATION

By default, the maximum notes per denomination when change-making for users

KIND

LOG_TARGET

MINT_E_CASH_TYPE_CHILD_ID 🔒

Traits

NotesSelector

Functions

select_notes_from_stream 🔒

spendable_notes_to_operation_id