Struct fedimint_mint_client::MintClientModule

pub struct MintClientModule {
    pub(crate) cfg: MintClientConfig,
    pub(crate) secret: DerivableSecret,
    pub(crate) secp: Secp256k1<All>,
    pub(crate) notifier: ModuleNotifier<DynGlobalClientContext, MintClientStateMachines>,
    pub(crate) cancel_oob_payment_bc: Sender<OperationId>,
}

Fields
cfg: MintClientConfigsecret: DerivableSecretsecp: Secp256k1<All>notifier: ModuleNotifier<DynGlobalClientContext, MintClientStateMachines>cancel_oob_payment_bc: Sender<OperationId>
Implementations

impl MintClientModule

pub async fn prepare_plaintext_ecash_backup(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
executor: Executor<DynGlobalClientContext>,
api: DynFederationApi,
module_instance_id: ModuleInstanceId
) -> Result<EcashBackup>

impl MintClientModule

pub async fn get_wallet_summary(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> TieredSummary

Returns the number of held e-cash notes per denomination

pub async fn create_output(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
operation_id: OperationId,
notes_per_denomination: u16,
amount: Amount
) -> ClientOutput<MintOutput, MintClientStateMachines>

Creates a mint output with exactly the given amount, issuing e-cash notes such that the client holds notes_per_denomination notes of each e-cash note denomination held.

pub async fn await_output_finalized(
&self,
operation_id: OperationId,
out_point: OutPoint
) -> Result<()>

Wait for the e-cash notes to be retrieved. If this is not possible because another terminal state was reached an error describing the failure is returned.

pub async fn create_input(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
operation_id: OperationId,
min_amount: Amount
) -> Result<ClientInput<MintInput, MintClientStateMachines>>

Creates a mint input of at least min_amount.

pub async fn create_input_from_notes(
&self,
operation_id: OperationId,
notes: TieredMulti<SpendableNote>
) -> Result<ClientInput<MintInput, MintClientStateMachines>>

Create a mint input from external, potentially untrusted notes

pub(crate) async fn spend_notes_oob(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
min_amount: Amount,
try_cancel_after: Duration
) -> Result<(OperationId, Vec<MintClientStateMachines>, TieredMulti<SpendableNote>)>

pub async fn await_spend_oob_refund(
&self,
operation_id: OperationId
) -> SpendOOBRefund

pub(crate) async fn await_restore_finished(&self) -> Result<()>

pub(crate) async fn select_notes(
dbtx: &mut ModuleDatabaseTransaction<'_>,
amount: Amount
) -> Result<TieredMulti<SpendableNote>, InsufficientBalanceError>

Select notes with total amount of at least amount. If more than requested amount of notes are returned it was because exact change couldn’t be made, and the next smallest amount will be returned.

The caller can request change from the federation.

pub(crate) async fn get_all_spendable_notes(
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> TieredMulti<SpendableNote>

pub(crate) async fn wipe_all_spendable_notes(
dbtx: &mut ModuleDatabaseTransaction<'_>
)

pub(crate) async fn get_next_note_index(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
amount: Amount
) -> NoteIndex

pub fn new_note_secret_static(
secret: &DerivableSecret,
amount: Amount,
note_idx: NoteIndex
) -> DerivableSecret

Derive the note DerivableSecret from the Mint’s secret the amount tier and note_idx

Static to help re-use in other places, that don’t have a whole Self available

pub(crate) async fn new_note_secret(
&self,
amount: Amount,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> DerivableSecret

pub async fn new_ecash_note(
&self,
amount: Amount,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> (NoteIssuanceRequest, BlindNonce)